APIrequirements.txt
fastapi==0.68.2
python-multipart==0.0.5
Pillow==8.4.0
azure-storage-blob==12.9.0
subprocess32==3.5.4
jinja2==3.0.3
uvicorn==0.15.0
